Added ShopVoucher

+++ ConsoleShoppingApp/src/DTO/Voucher/ItemVoucher.java
@@ -1,7 +1,7 @@
 package DTO.Voucher;
 
 public class ItemVoucher extends Voucher {
-    public ItemVoucher(float discountRate, int amount, float minimumSpend) {
+    public ItemVoucher(float discountRate, int amount, int minimumSpend) {
         super(discountRate, amount, minimumSpend);
         super.type = "ItemVoucher";
     }
