+++ ConsoleShoppingApp/src/DTO/Voucher/ShippingVoucher.java
@@ -1,7 +1,7 @@
 package DTO.Voucher;
 
 public class ShippingVoucher extends Voucher{
-    public ShippingVoucher(float discountRate, int amount, float minimumSpend) {
+    public ShippingVoucher(float discountRate, int amount, int minimumSpend) {
         super(discountRate, amount, minimumSpend);
         super.type = "ShippingVoucher";
     }
